% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{addAttribute}
\alias{addAttribute}
\title{Add an attribute to an object.}
\usage{
addAttribute(object, name, value)
}
\arguments{
\item{object}{The object for which we want modification.}

\item{name}{The name of the attribute we want to add.}

\item{value}{The value we want to give to the attribute.}
}
\value{
The object without the attribute you want to delete.
}
\description{
Add an attribute to an object.
}
