% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/addAttribute-methods.R
\docType{methods}
\name{addAttribute}
\alias{addAttribute}
\alias{addAttribute}
\alias{addAttribute,Experiment-method}
\alias{addAttribute}
\alias{addAttribute,Gene-method}
\title{Add an attribute to an object.}
\usage{
addAttribute(object, name, value)

\S4method{addAttribute}{Experiment}(object, name, value)

\S4method{addAttribute}{Gene}(object, name, value)
}
\arguments{
\item{object}{The object for which we want modification.}

\item{name}{The name of the attribute we want to add.}

\item{value}{The value we want to give to the attribute.}
}
\value{
The object without the attribute you want to delete.
}
\description{
Add an attribute to an object.
}
\examples{
exp <- new(Class="Experiment",
name="test",
date=Sys.Date(),
databases=list(1),
others=list())
exp <- addAttribute(exp, "test", "value")
gene <- new("RAPDB")
gene <- addAttribute(gene, "test", "value")
}
