% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pratical-functions.R
\name{alreadyUsedDB}
\alias{alreadyUsedDB}
\title{Function checking if the database is already used}
\usage{
alreadyUsedDB(databases, i)
}
\arguments{
\item{databases}{the list of the databases we have}

\item{i}{the number of the database we want to know if it's already used}
}
\value{
Boolean TRUE if the database is already used, FALSE if not
}
\description{
Function checking if the database is already used
}
