% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{analyseDataDB7}
\alias{analyseDataDB7}
\title{analyseDataDB7}
\usage{
analyseDataDB7(rOutput)
}
\arguments{
\item{rOutput}{character}
}
\value{
it will return ...
}
\description{
This function is called
}
