% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CallDB.R
\name{Appel_Scriptv7}
\alias{Appel_Scriptv7}
\title{Fonction qui appel le Scriptv7 de python}
\usage{
Appel_Scriptv7(RAPID)
}
\description{
Cette fonction permet à l'utilisateur de chercher si le RAPID existe dans Oryzabase.txt
}
\examples{
Appel_Scriptv7("Os06g0654600")
}
