% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{callCreationGeneDB10}
\alias{callCreationGeneDB10}
\title{callCreationGeneDB10}
\usage{
callCreationGeneDB10(x, IdsList, locusList)
}
\arguments{
\item{x}{number}

\item{IdsList}{list}

\item{locusList}{list}
}
\value{
It will return a list of genes
}
\description{
This function calls creationGeneDB10 function
}
