% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{callCreationGeneDB5}
\alias{callCreationGeneDB5}
\title{callCreationGeneDB5}
\usage{
callCreationGeneDB5(x, IdsList, locusList)
}
\arguments{
\item{x}{number}

\item{IdsList}{list}

\item{locusList}{list}
}
\value{
It will return a list of genes
}
\description{
This function calls creationGeneDB5 function
}
