% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{callCreationGeneDB6}
\alias{callCreationGeneDB6}
\title{callCreationGeneDB6}
\usage{
callCreationGeneDB6(x, IdsList, locusList)
}
\arguments{
\item{x}{number}

\item{IdsList}{list}

\item{locusList}{list}
}
\value{
It will return a list of genes
}
\description{
This function calls creationGeneDB6 function
}
