% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{callCreationGeneDB7}
\alias{callCreationGeneDB7}
\title{callCreationGeneDB7}
\usage{
callCreationGeneDB7(x, IdsList, locusList)
}
\arguments{
\item{x}{number}

\item{IdsList}{list}

\item{locusList}{list}
}
\value{
It will return a list of genes
}
\description{
This function calls creationGeneDB7 function
}
