% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{callCreationGeneDB9}
\alias{callCreationGeneDB9}
\title{callCreationGeneDB9}
\usage{
callCreationGeneDB9(x, IdsList, locusList)
}
\arguments{
\item{x}{number}

\item{IdsList}{list}

\item{locusList}{list}
}
\value{
It will return a list of genes
}
\description{
This function calls creationGeneDB9 function
}
