% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CallDB.R
\name{callDB1}
\alias{callDB1}
\title{callDB1}
\usage{
callDB1(IdsList, locusList)
}
\arguments{
\item{IdsList}{list of locus for which we want the genes}

\item{locusList}{list}
}
\description{
This function is the new function which will replace callDB1 function
}
