% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CallDB.R
\name{CallDB1}
\alias{CallDB1}
\title{This function calls run.py}
\usage{
CallDB1(locusListe)
}
\arguments{
\item{Locus}{list}
}
\description{
This function will call for each locus in the list, the run.py script and
python will return the list of the genes which are present in the locus and
in the DB1. All these locus will be stocked in liste_genes
}
\examples{
callDB1(Liste de locus)
}
