% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{callDB2}
\alias{callDB2}
\title{callDB2}
\usage{
callDB2(IdsList, locusList)
}
\arguments{
\item{IdsList}{list of locus for which we want the genes}

\item{locusList}{list}
}
\value{
It will return only a list with all the genesDB2
}
\description{
This function will call the Gramene database
}
