% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{callDB3}
\alias{callDB3}
\title{callDB3}
\usage{
callDB3(IdsList, locusList)
}
\arguments{
\item{IdsList}{list of locus for which we want the genes}

\item{locusList}{list}
}
\value{
It will return only a list with all the genesDB3
}
\description{
This function will call the third database
}
