% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{callDB9}
\alias{callDB9}
\title{callDB9}
\usage{
callDB9(IdsList, locusList)
}
\arguments{
\item{IdsList}{list of locus for which we want the genes}

\item{locusList}{list}
}
\value{
It will return only a list with all the genesDB9
}
\description{
This function will call
}
\examples{
locusList <- data.frame(ch = c("1","1"),
                        st = c("148907","527906"),
                        end = c("248907","842359"))
                        
ids <- list(list("Os01g0102700","Os01g0102800"),
            list("Os01g0109750","Os01g0110100"))
                
callDB9(ids, locusList)
}
