% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-linkPython.R
\name{callExecutablePy}
\alias{callExecutablePy}
\title{callExecutablePy}
\usage{
callExecutablePy()
}
\description{
This function calls test.py which will be allow to all python file to be
executable
}
