% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callSnpSeek-function.R
\name{callSnpSeek}
\alias{callSnpSeek}
\title{Function for recup a list of ids}
\usage{
callSnpSeek(locus)
}
\arguments{
\item{locus}{the list of locus for which we want the ids}
}
\value{
the list of id's of each genes we want
}
\description{
Function for recup a list of ids
}
\examples{
locusList <- data.frame(ch = c("1","1"),
                        st = c("148907","527906"),
                        end = c("248907","842359"))
                  
callSnpSeek(locusList)
}
