% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInformations-functions.R
\name{changeNumberIntoDBName}
\alias{changeNumberIntoDBName}
\title{Function to know the the name of the database associate to the number}
\usage{
changeNumberIntoDBName(number)
}
\value{
return the name of the database associate to the number
}
\description{
this function return the name of the database associate to the number
}
