% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pratical-functions.R
\name{createAttributesVector}
\alias{createAttributesVector}
\title{Function to create the vector of the attributes you want from a class}
\usage{
createAttributesVector(class, attributesVector)
}
\arguments{
\item{class}{the name of the class you want}

\item{attributesVector}{the attributes you want to extract from the class}
}
\value{
print the attributes of the class
}
\description{
this function will give you the vector to use selectProperties
}
\examples{
createAttributesVector("RAPDB",c("id"))
}
