% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{creationGeneDB10}
\alias{creationGeneDB10}
\title{creationGeneDB10}
\usage{
creationGeneDB10(x, y, IdsList, locusList)
}
\arguments{
\item{x}{number}

\item{y}{number}

\item{IdsList}{list}

\item{locusList}{list}
}
\value{
it will return a GeneDB10
}
\description{
This function is called only by callDB10 and will create the gene DB10
It will call run.py script which will return the list of the genes which
are present in the locus
}
