% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{creationGeneDB2}
\alias{creationGeneDB2}
\title{creationGeneDB2}
\usage{
creationGeneDB2(x, y, IdsList, locusList)
}
\arguments{
\item{x}{number}

\item{y}{number}

\item{IdsList}{list}

\item{locusList}{list}
}
\value{
it will return a GeneDB2
}
\description{
This function is called only by callDB1 and will create the gene DB2
It will call run.py script which will return the list of the genes which
are present in the locus
}
