% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CallDB.R
\name{creationGeneDB3}
\alias{creationGeneDB3}
\title{creationGeneDB3}
\usage{
creationGeneDB3(i, locusList)
}
\arguments{
\item{i}{number}

\item{locusList}{list}
}
\value{
it will return a geneDB3
}
\description{
This function is called only by callDB3 and will create the gene DB3
It will call run.py script which will return the list of the genes which 
are present in the locus
}
