% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{creationGeneDB4}
\alias{creationGeneDB4}
\title{creationGeneDB4}
\usage{
creationGeneDB4(x, y, IdsList, locusList)
}
\arguments{
\item{x}{number}

\item{y}{number}

\item{IdsList}{list}

\item{locusList}{list}
}
\value{
it will return a GeneDB4
}
\description{
This function is called only by callDB4 and will create the gene DB4
It will call run.py script which will return the list of the genes which
are present in the locus
}
