% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{creationGeneDB5}
\alias{creationGeneDB5}
\title{creationGeneDB5}
\usage{
creationGeneDB5(x, y, IdsList, locusList)
}
\arguments{
\item{x}{number}

\item{y}{number}

\item{IdsList}{list}

\item{locusList}{list}
}
\value{
it will return a GeneDB5
}
\description{
This function is called only by callDB8 and will create the gene DB5
It will call run.py script which will return the list of the genes which
are present in the locus
}
