% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{creationGeneDB6}
\alias{creationGeneDB6}
\title{creationGeneDB6}
\usage{
creationGeneDB6(x, y, IdsList, locusList)
}
\arguments{
\item{x}{number}

\item{y}{number}

\item{IdsList}{list}

\item{locusList}{list}
}
\value{
it will return a GeneDB6
}
\description{
This function is called only by callDB6 and will create the gene DB6
It will call run.py script which will return the list of the genes which
are present in the locus
}
