% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{creationGeneDB7}
\alias{creationGeneDB7}
\title{creationGeneDB7}
\usage{
creationGeneDB7(x, y, IdsList, locusList)
}
\arguments{
\item{x}{number}

\item{y}{number}

\item{IdsList}{list}

\item{locusList}{list}
}
\value{
it will return a GeneDB7
}
\description{
This function is called only by callDB7 and will create the gene DB7
It will call run.py script which will return the list of the genes which
are present in the locus
}
