% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{creationGeneDB9}
\alias{creationGeneDB9}
\title{creationGeneDB9}
\usage{
creationGeneDB9(x, y, IdsList, locusList)
}
\arguments{
\item{x}{number}

\item{y}{number}

\item{IdsList}{list}

\item{locusList}{list}
}
\value{
it will return a GeneDB9
}
\description{
This function is called only by callDB9 and will create the gene DB9
It will call run.py script which will return the list of the genes which
are present in the locus
}
