% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInformations-functions.R
\name{databasesAvailables}
\alias{databasesAvailables}
\title{Function to know the number of databases availables}
\usage{
databasesAvailables()
}
\value{
return the number of databases availables
}
\description{
this function return the number of databases availables
}
\examples{
databasesAvailables()
}
