% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInformations-functions.R
\name{databasesList}
\alias{databasesList}
\title{Function for see the list of the databases we can call}
\usage{
databasesList()
}
\value{
print the databases we can call
}
\description{
this function will print the databases we can call
}
\examples{
databasesList()
}
