% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{deleteAttribute}
\alias{deleteAttribute}
\title{Delete an attribute which has been add by the user.}
\usage{
deleteAttribute(object, name)
}
\arguments{
\item{object}{The object for which we want modification.}

\item{name}{The name of the attribute we want to delete.}
}
\value{
The object without the attribute you want to delete.
}
\description{
Delete an attribute which has been add by the user.
}
