% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pratical-functions.R
\name{existsGene}
\alias{existsGene}
\title{Function that allows to know if a gene already exists or not}
\usage{
existsGene(genes, id)
}
\arguments{
\item{genes}{the list of unique genes.}

\item{id}{the id of the gene we don't want to duplicate.}
}
\value{
Boolean TRUE if the gene exists, FALSE if not.
}
\description{
Function that allows to know if a gene already exists or not
}
