% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/methods-Experiment.R
\docType{classes}
\name{Experiment}
\alias{Experiment}
\alias{deleteAttribute,Experiment-method}
\title{The Experiment class.}
\usage{
\S4method{deleteAttribute}{Experiment}(object, name)
}
\description{
Some details about this class and my plans for it in the body.
}
\section{Slots}{

\describe{
\item{\code{name}}{The name of the experiment}

\item{\code{date}}{The date of the experiment}

\item{\code{databases}}{the databases which concern the object we are studying}

\item{\code{genes}}{The list of genes we avec download.}

\item{\code{others}}{The others attributes the user can add and delete}
}}

