% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors-functions.R
\name{Experiment}
\alias{Experiment}
\title{Constructor of an Experiment}
\usage{
Experiment(name, date, databases, genes)
}
\arguments{
\item{name}{the name of the experiment}

\item{date}{the date of the experiment}

\item{databases}{the list of the databases}

\item{genes}{the list of genes of the databases}
}
\description{
allows to create an experiment
}
