% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Funricigenes2-class}
\alias{Funricigenes2-class}
\title{An S4 class to represent the Gene database Funricigenes 2}
\description{
This gene is specific to the database Funricigenes 2. So it
contains all the informations we can have about one gene with the
"Funricigenes 2" database.
}
\section{Slots}{

\describe{
\item{\code{id}}{The id of the gene. Many genes can have the same id because 
there are the same in spite of the different attributes they have.}

\item{\code{genesIDs}}{the list of diferent ids it have we will use with the 
databases.}

\item{\code{locus}}{The locus the gene belongs.}

\item{\code{others}}{The others attributes the user want to add or delete.}

\item{\code{symbol}}{the gene symbol}

\item{\code{name}}{name}
}}

\examples{
gene <- new("Funricigenes2")
}
