% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Gene}
\alias{Gene}
\title{The Gene class.}
\description{
A gene contains a lot of details that depend on the database it belongs. 
But every genes have informations in common. So we put all these informations
into a global class "gene" from which the others specific gene class will
inherit.
}
\section{Slots}{

\describe{
\item{\code{id}}{The id of the gene. Many genes can have the same id because there
are the same in spite of the different attributes they have.}

\item{\code{locus}}{The locus the gene belongs.}

\item{\code{others}}{The others attributes the user want to add or delete.}

\item{\code{genes}}{The list of genes we avec download.}
}}

