% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GeneDB3}
\alias{GeneDB3}
\title{The Gene  class specific to the third database.}
\description{
This gene is specific to the database 3 that is the database "Oryzabase". 
So it contains all the informations we can have about one gene with the 
"Oryzabase" database.
}
\section{Slots}{

\describe{
\item{\code{locus}}{}

\item{\code{GeneId}}{}
}}

