% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getAttributes-methods.R
\docType{methods}
\name{getAttributesNames}
\alias{getAttributesNames}
\alias{getAttributesNames,Experiment-method}
\alias{getAttributesNames,Gene-method}
\title{Get the attributes of an object.}
\usage{
getAttributesNames(object, nbDB)

\S4method{getAttributesNames}{Experiment}(object)

\S4method{getAttributesNames}{Gene}(object)
}
\arguments{
\item{object}{The object for which we want to know the attributes.}
}
\value{
the list of the attributes wanted.
}
\description{
Get the attributes of an object.
}
