% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{getDate}
\alias{getDate}
\title{Get the date of an object}
\usage{
getDate(object)
}
\arguments{
\item{object}{The object for which we want to get the name}
}
\value{
The date of the object
}
\description{
Get the date of an object
}
