% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getDBGenes-methods.R
\docType{methods}
\name{getDBGenes}
\alias{getDBGenes}
\alias{getDBGenes,Experiment-method}
\title{Get all the genes of the data base you want.}
\usage{
getDBGenes(object, nbDB)

\S4method{getDBGenes}{Experiment}(object, nbDB)
}
\arguments{
\item{object}{The object for which we want modification.}

\item{nbDB}{The number of the db you want(ex : 1 for RAPDB).}
}
\value{
the list of the genes of the database you want.
}
\description{
Get all the genes of the data base you want.
}
