% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/get-methods.R
\docType{methods}
\name{getGenes}
\alias{getGenes}
\alias{getGenes,Experiment-method}
\title{Get the genes of an object}
\usage{
getGenes(object)

\S4method{getGenes}{Experiment}(object)
}
\arguments{
\item{object}{The object for which we want to get the genes}
}
\value{
The genes of the object
}
\description{
Get the genes of an object
}
\examples{
gene <- new("Gene", id="exId", locus=data.frame(1,222,333) )

exp <- new(Class="Experiment",
 date=Sys.Date(),
  name="ex",
  genes=list(list(gene)),
   databases=list(1))

getGenes(exp)
}
