% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/get-methods.R
\docType{methods}
\name{getId}
\alias{getId}
\alias{getId,Gene-method}
\title{Get the id of an object}
\usage{
getId(object)

\S4method{getId}{Gene}(object)
}
\arguments{
\item{object}{The object for which we want to get the id}
}
\value{
The id of the object
}
\description{
Get the id of an object
}
\examples{
gene <- new("Gene", id="exId" )

getId(gene)
}
