% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callSnpSeek-function.R
\name{getIds}
\alias{getIds}
\title{getIds}
\usage{
getIds(i, locusList)
}
\arguments{
\item{i}{number}

\item{locusList}{list}
}
\description{
This function is called for each locus and has to return the list of ids 
present in the locus
}
