% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/get-methods.R
\docType{methods}
\name{getLocus}
\alias{getLocus}
\alias{getLocus,Gene-method}
\title{Get the locus of an object}
\usage{
getLocus(object)

\S4method{getLocus}{Gene}(object)
}
\arguments{
\item{object}{The object for which we want to get the locus}
}
\value{
The locus of the object
}
\description{
Get the locus of an object
}
\examples{
gene <- new("Gene", id="exId", locus=data.frame(1,222,333) )

getLocus(gene)
}
