% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{getName}
\alias{getName}
\title{Get the name of an object}
\usage{
getName(object)
}
\arguments{
\item{object}{The object for which we want to get the name}
}
\value{
The name of the object
}
\description{
Get the name of an object
}
