% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pratical-functions.R
\name{getOutPutJSON}
\alias{getOutPutJSON}
\title{Function for a JSON return}
\usage{
getOutPutJSON(outPut)
}
\arguments{
\item{outPut}{character}
}
\value{
return only string which starts with "\{" -> JSON
}
\description{
this function will only return an JSON return which start with "\{"
It will allow us to treat the exception error from python
}
