% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Gramene-class}
\alias{Gramene-class}
\title{An S4 class to represent the Gene database Gramene}
\description{
This gene is specific to the database Gramene. So it
contains all the informations we can have about one gene with the
"Gramene" database.
}
\section{Slots}{

\describe{
\item{\code{id}}{The id of the gene. Many genes can have the same id because 
there are the same in spite of the different attributes they have.}

\item{\code{genesIDs}}{the list of diferent ids it have we will use with the 
databases.}

\item{\code{locus}}{The locus the gene belongs.}

\item{\code{others}}{The others attributes the user want to add or delete.}

\item{\code{description}}{description of the gene}

\item{\code{biotype}}{the biotype of the gene}

\item{\code{taxonId}}{id}

\item{\code{systemName}}{system name}

\item{\code{dbType}}{type of the db}

\item{\code{geneIdx}}{id}

\item{\code{location}}{location}
}}

\examples{
gene <- new("Gramene")
}
