% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{IC4R-class}
\alias{IC4R-class}
\title{An S4 class to represent Gene database IC4R.}
\description{
This gene is specific to the database "IC4R". So it
contains all the informations we can have about one gene with the "IC4R"
database.
}
\section{Slots}{

\describe{
\item{\code{id}}{The id of the gene. Many genes can have the same id because 
there are the same in spite of the different attributes they have.}

\item{\code{genesIDs}}{the list of diferent ids it have we will use with the 
databases.}

\item{\code{locus}}{The locus the gene belongs.}

\item{\code{others}}{The others attributes the user want to add or delete.}

\item{\code{experimentName}}{the experiment name}

\item{\code{devStage}}{devStage}

\item{\code{experimentId}}{the experiment id,}

\item{\code{variety}}{the variety,}

\item{\code{projectId}}{the project id,}

\item{\code{tissue}}{the tissue,}

\item{\code{exprValue}}{the expr value}
}}

\examples{
gene <- new("IC4R")
}
