% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callSnpSeek-function.R
\name{id}
\alias{id}
\title{id}
\usage{
id(rOutput)
}
\arguments{
\item{rOutput}{character}
}
\value{
this funciton will only return the id
}
\description{
This function will only return the id from the rOutput
}
