% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbInformations-functions.R
\name{listDB}
\alias{listDB}
\title{Function to know the list of the databases availables}
\usage{
listDB()
}
\value{
return a list of the databases availables
}
\description{
this function return return a list of the databases availables
}
