% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{makeList}
\alias{makeList}
\title{makeList}
\usage{
makeList(liste, number)
}
\arguments{
\item{liste}{list}

\item{number}{number}
}
\value{
it will return ...
}
\description{
This function is called in order to return element in order to restructure
a list
}
