% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{PLANTTFDB-class}
\alias{PLANTTFDB-class}
\title{An S4 class to represent the Gene database PLANTTFDB}
\description{
This gene is specific to the database PLANTTFDB So it
contains all the informations we can have about one gene with the
"PLANTTFDB" database.
}
\section{Slots}{

\describe{
\item{\code{id}}{The id of the gene. Many genes can have the same id because 
there are the same in spite of the different attributes they have.}

\item{\code{genesIDs}}{the list of diferent ids it have we will use with the 
databases.}

\item{\code{locus}}{The locus the gene belongs.}

\item{\code{others}}{The others attributes the user want to add or delete.}

\item{\code{family}}{family}
}}

\examples{
gene <- new("PLANTTFDB")
}
