% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RAPDB-class}
\alias{RAPDB-class}
\title{An S4 class to represent the first database Gene.}
\description{
This gene is specific to the database 1 that is the database "RAPDB". So it
contains all the informations we can have about one gene with the "RAPDB"
database.
}
\section{Slots}{

\describe{
\item{\code{id}}{The id of the gene. Many genes can have the same id because 
there are the same in spite of the different attributes they have.}

\item{\code{genesIDs}}{the list of diferent ids it have we will use with the 
databases.}

\item{\code{locus}}{The locus the gene belongs.}

\item{\code{others}}{The others attributes the user want to add or delete.}

\item{\code{rapDBGeneNameSynonym}}{an identificator}

\item{\code{rapDBGeneSymbolSynonym}}{an identificator}

\item{\code{cgsnlGeneName}}{an identificator}

\item{\code{cgsnlGeneSymbol}}{an identificator}

\item{\code{oryzabaseGeneNameSynonym}}{an identificator}

\item{\code{oryzabaseGeneSymbolSynonym}}{an identificator}

\item{\code{position}}{the position of the gene into the chromosome}

\item{\code{description}}{a description of the gene}
}}

\examples{
gene <- new("RAPDB")
}
