% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pratical-functions.R
\name{returnError}
\alias{returnError}
\title{Function which returns only the right errors}
\usage{
returnError(outPut)
}
\arguments{
\item{outPut}{character}
}
\value{
return only errors we know
}
\description{
This function will return all the errors like "Bad request",...
}
