% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/selectProperties-methods.R
\docType{methods}
\name{selectProperties}
\alias{selectProperties}
\alias{selectProperties,Experiment-method}
\title{select the properties you want to study}
\usage{
selectProperties(object, attributesVector)

\S4method{selectProperties}{Experiment}(object, attributesVector)
}
\arguments{
\item{object}{The object for which we want to extract a dataframe}

\item{attributesVector}{list of the attributes you want to keep}
}
\value{
a data frame with the genes and the properties you want
}
\description{
select the properties you want to study
}
