% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{setDate}
\alias{setDate}
\title{Set the date of an object}
\usage{
setDate(object, date)
}
\arguments{
\item{object}{The object for which we want to set the name}

\item{date}{The date we want to give at the object}
}
\value{
The object with the new name
}
\description{
Set the date of an object
}
