% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set-methods.R
\docType{methods}
\name{setName}
\alias{setName}
\alias{setName,Experiment-method}
\title{Set the name of an object}
\usage{
setName(object, name)

\S4method{setName}{Experiment}(object, name)
}
\arguments{
\item{object}{The object for which we want to set the name}

\item{name}{The name we want to give at the object}
}
\value{
The object with the new name

The object with the new name
}
\description{
Set the name of an object
}
\examples{
exp <- new(Class="Experiment",
 databases=list(1),
 date=Sys.Date(),
 name="test")
 
exp <- setName(exp, "example")
}
