% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{setName}
\alias{setName}
\title{Set the name of an object}
\usage{
setName(object, name)
}
\arguments{
\item{object}{The object for which we want to set the name}

\item{name}{The name we want to give at the object}
}
\value{
The object with the new name
}
\description{
Set the name of an object
}
