% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/updateObject-methods.R
\docType{methods}
\name{updateObject}
\alias{updateObject}
\alias{updateObject,Experiment-method}
\title{Update an attribute of an object}
\usage{
updateObject(object, attribute, value)

\S4method{updateObject}{Experiment}(object, attribute, value)
}
\arguments{
\item{object}{The object for which we want modification}

\item{attribute}{The attribute we want to update}

\item{value}{The value we want to give at the attribute}
}
\value{
The object without the attribute we want to delete
}
\description{
Update an attribute of an object
}
