% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pratical-functions.R
\name{whichOS}
\alias{whichOS}
\title{Function to know the OS using the package}
\usage{
whichOS()
}
\value{
numeric 0 if windows, 1 if linux
}
\description{
this function allows us to know if the user is using Linux or Windows
}
