% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{analyseDataDB4}
\alias{analyseDataDB4}
\title{analyseDataDB7}
\usage{
analyseDataDB4(rOutput)
}
\arguments{
\item{rOutput}{character}
}
\value{
it will return a list of elements
}
\description{
This function returns a list with importants elements
}
