% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CallDB.R
\name{appel_Scriptv7}
\alias{appel_Scriptv7}
\title{This function calls Scriptv7 from python}
\usage{
appel_Scriptv7(RAPID)
}
\description{
Cette fonction permet à l'utilisateur de chercher si le RAPID existe dans 
Oryzabase.txt
}
\examples{
Appel_Scriptv7("Os06g0654600")
}
