% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CallDB.R
\name{callDB1}
\alias{callDB1}
\title{callDB1}
\usage{
callDB1(IdsList, locusList)
}
\arguments{
\item{IdsList}{list of locus for which we want the genes}

\item{locusList}{list}
}
\description{
This function is the new function which will replace callDB1 function
}
\examples{
locusList <- data.frame(ch = c("1","1"),
                        st = c("148907","527906"),
                        end = c("248907","842359"))
                        
ids <- list(list("Os01g0102700","Os01g0102800"),
            list("Os01g0109750","Os01g0110100"))
                
callDB1(ids, locusList)
}
