% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{callDB1}
\alias{callDB1}
\title{callDB1}
\usage{
callDB1(IdsList, locusList)
}
\arguments{
\item{IdsList}{list of ids and uniquename we can catch with callSnpSeek}

\item{locusList}{list of locus for which we want the genes}
}
\value{
It will return only a list with all the genesDB1
}
\description{
This function is linked to RAPDB database and will return a list of genes
}
\examples{
locusList <- data.frame(ch = c("1"),
                        st = c("148907"),
                        end = c("248907"))
                        
ids <- list(list("testId"))
                
callDB1(ids, locusList)
}
