% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{callDB2}
\alias{callDB2}
\title{callDB2}
\usage{
callDB2(IdsList, locusList)
}
\arguments{
\item{IdsList}{list of ids and uniquename we can catch with callSnpSeek}

\item{locusList}{list of locus for which we want the genes}
}
\value{
It will return only a list with all the genesDB2
}
\description{
This function is linked to GRAMENE database and will return a list of genes
}
\examples{
locusList <- data.frame(ch = c("1"),
                        st = c("148907"),
                        end = c("248907"))
                        
ids <- list(list("testId"))
                
callDB2(ids, locusList)
}
