% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CallDB.R
\name{callDB3}
\alias{callDB3}
\title{callDB3}
\usage{
callDB3(locusList)
}
\arguments{
\item{locusList}{list}
}
\description{
This function will call for each locus in the list, the run.py script and
python will return the list of the genes which are present in the locus and
in the DB3. All these locus will be stocked in listGenes
}
