% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{callDB3}
\alias{callDB3}
\title{callDB3}
\usage{
callDB3(IdsList, locusList)
}
\arguments{
\item{IdsList}{list of ids and uniquename we can catch with callSnpSeek}

\item{locusList}{list of locus for which we want the genes}
}
\value{
It will return only a list with all the genesDB3
}
\description{
This function is linked to ORYZABASE database and will return a list of 
genes
}
\examples{
locusList <- data.frame(ch = c("1","1"),
                        st = c("148907","527906"),
                        end = c("248907","842359"))
                        
ids <- list(list("testId"))
                
callDB3(ids, locusList)
}
