% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callSnpSeek-function.R
\name{callSnpSeek}
\alias{callSnpSeek}
\title{CallSnpSeek}
\usage{
callSnpSeek(locus)
}
\arguments{
\item{locus}{the list of locus for which we want the ids}
}
\value{
the list of id's of each genes we want
}
\description{
It will return a list of IDs and msu7Name
}
\examples{
locusList <- data.frame(ch = c("1"),
                        st = c("148907"),
                        end = c("248907"))
                  
callSnpSeek(locusList)
}
