% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{creationGeneDB1}
\alias{creationGeneDB1}
\title{creationGeneDB1}
\usage{
creationGeneDB1(x, y, IdsList, locusList)
}
\arguments{
\item{x}{number}

\item{y}{number}

\item{IdsList}{list}

\item{locusList}{list}
}
\value{
it will return a gene from the class RAPDB
}
\description{
This function returns a gene specific to the RAPDB database
}
