% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{creationGeneDB10}
\alias{creationGeneDB10}
\title{creationGeneDB10}
\usage{
creationGeneDB10(x, y, IdsList, locusList)
}
\arguments{
\item{x}{number}

\item{y}{number}

\item{IdsList}{list}

\item{locusList}{list}
}
\value{
it will return a gene from the class MSU
}
\description{
This function returns a gene specific to the MSU database
}
