% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{creationGeneDB2}
\alias{creationGeneDB2}
\title{creationGeneDB2}
\usage{
creationGeneDB2(x, y, IdsList, locusList)
}
\arguments{
\item{x}{number}

\item{y}{number}

\item{IdsList}{list}

\item{locusList}{list}
}
\value{
it will return a gene from the class Gramene
}
\description{
This function returns a gene specific to the GRAMENE database
}
