% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{creationGeneDB3}
\alias{creationGeneDB3}
\title{creationGeneDB3}
\usage{
creationGeneDB3(x, y, IdsList, locusList)
}
\arguments{
\item{x}{number}

\item{y}{number}

\item{IdsList}{list}

\item{locusList}{list}
}
\value{
it will return a gene from the class Oryzabase
}
\description{
This function returns a gene specific to the ORYZABASE database
}
