% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{creationGeneDB4}
\alias{creationGeneDB4}
\title{creationGeneDB4}
\usage{
creationGeneDB4(x, y, IdsList, locusList)
}
\arguments{
\item{x}{number}

\item{y}{number}

\item{IdsList}{list}

\item{locusList}{list}
}
\value{
it will return a gene from the class IC4R
}
\description{
This function returns a gene specific to the IC4R database
}
