% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{creationGeneDB5}
\alias{creationGeneDB5}
\title{creationGeneDB5}
\usage{
creationGeneDB5(x, y, IdsList, locusList)
}
\arguments{
\item{x}{number}

\item{y}{number}

\item{IdsList}{list}

\item{locusList}{list}
}
\value{
it will return a gene from the class PLANTTFDB
}
\description{
This function returns a gene specific to the PLANTTFDB database
}
