% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{creationGeneDB6}
\alias{creationGeneDB6}
\title{creationGeneDB6}
\usage{
creationGeneDB6(x, y, IdsList, locusList)
}
\arguments{
\item{x}{number}

\item{y}{number}

\item{IdsList}{list}

\item{locusList}{list}
}
\value{
it will return a gene from the class PLNTFDB
}
\description{
This function returns a gene specific to the PLNTFDB database
}
