% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{creationGeneDB7}
\alias{creationGeneDB7}
\title{creationGeneDB7}
\usage{
creationGeneDB7(x, y, IdsList, locusList)
}
\arguments{
\item{x}{number}

\item{y}{number}

\item{IdsList}{list}

\item{locusList}{list}
}
\value{
it will return a gene from the class Funricigenes
}
\description{
This function returns a gene specific to the FUNRICIGENES database
}
