% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{creationGeneDB8}
\alias{creationGeneDB8}
\title{creationGeneDB8}
\usage{
creationGeneDB8(x, y, IdsList, locusList)
}
\arguments{
\item{x}{number}

\item{y}{number}

\item{IdsList}{list}

\item{locusList}{list}
}
\value{
it will return a gene from the class Funricigenes2
}
\description{
This function returns a gene specific to the FUNRICIGENES2 database
}
