% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callDB-functions.R
\name{creationGeneDB9}
\alias{creationGeneDB9}
\title{creationGeneDB9}
\usage{
creationGeneDB9(x, y, IdsList, locusList)
}
\arguments{
\item{x}{number}

\item{y}{number}

\item{IdsList}{list}

\item{locusList}{list}
}
\value{
it will return a gene from the class Funricigenes3
}
\description{
This function returns a gene specific to the FUNRICIGENES3 database
}
