% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/deleteAttribute-methods.R
\docType{methods}
\name{deleteAttribute}
\alias{deleteAttribute}
\alias{deleteAttribute,Experiment-method}
\alias{deleteAttribute}
\alias{deleteAttribute,Gene-method}
\title{Delete an attribute which has been add by the user.}
\usage{
deleteAttribute(object, name)

\S4method{deleteAttribute}{Experiment}(object, name)

\S4method{deleteAttribute}{Gene}(object, name)
}
\arguments{
\item{object}{The object for which we want modification.}

\item{name}{The name of the attribute we want to delete.}
}
\value{
The object without the attribute you want to delete.
}
\description{
Delete an attribute which has been add by the user.
}
\examples{
exp <- new(Class="Experiment",
name="test",
date=Sys.Date(),
databases=list(1),
others=list())
exp <- addAttribute(exp, "test", "value")
exp <- deleteAttribute(exp, "test")
gene <- new("RAPDB")
gene <- addAttribute(gene, "test", "value")
gene <- deleteAttribute(gene, "test")
}
