% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkPython-functions.R
\name{dlPythonModules}
\alias{dlPythonModules}
\title{dlPythonModules}
\usage{
dlPythonModules()
}
\value{
nothing
}
\description{
This function calls importModules.py in order to install all the python
modules we need if they have not yet been installed
}
