% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/methods-Experiment.R
\docType{class}
\name{Experiment}
\alias{Experiment}
\alias{Experiment}
\alias{addAttribute,Experiment-method}
\alias{updateObject,Experiment-method}
\alias{deleteAttribute,Experiment-method}
\title{The Experiment class.}
\usage{
Experiment(name, locus)

\S4method{updateObject}{Experiment}(object, attribute, value)

\S4method{deleteAttribute}{Experiment}(object, name)
}
\arguments{
\item{name}{The name of the Experiment}

\item{locus}{The Table of locus}
}
\value{
An Experiment
}
\description{
Some details about this class and my plans for it in the body.

This constructor allows to create an Experiment object, with a name and a
locus list. The locus list is a .txt file, so you have to do a "read.table"
of your file into a variable and put this variable as the "locus" parameter
in the constructor.
}
\section{Slots}{

\describe{
\item{\code{name}}{The name of the experiment}

\item{\code{date}}{The date of the experiment}

\item{\code{databases}}{the databases which concern the object we are studying}

\item{\code{genes}}{The list of genes we avec download.}

\item{\code{others}}{The others attributes the user can add and delete}
}}

