% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Experiment-class}
\alias{Experiment-class}
\title{An S4 class to represent an Experiment class.}
\description{
Some details about this class and my plans for it in the body.
}
\section{Slots}{

\describe{
\item{\code{name}}{The name of the experiment}

\item{\code{date}}{The date of the experiment}

\item{\code{databases}}{the databases which concern the object we are studying}

\item{\code{genes}}{The list of genes we avec download.}

\item{\code{properties}}{The properties you want to keep}

\item{\code{others}}{The others attributes the user can add and delete}
}}

\examples{
exp <- new(Class="Experiment", date=Sys.Date(), name="test")
}
