% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Gene-class}
\alias{Gene-class}
\title{An S4 class to represent a Gene.}
\description{
A gene contains a lot of details that depend on the database it belongs.
But every genes have informations in common. So we put all these informations
into a global class "Gene" from which the others specific genes classes will
inherit.
}
\section{Slots}{

\describe{
\item{\code{id}}{The id of the gene. Many genes can have the same id because 
there are the same in spite of the different attributes they have.}

\item{\code{genesIDs}}{the list of diferent ids it have we will use with the 
databases.}

\item{\code{locus}}{The locus the gene belongs.}

\item{\code{others}}{The others attributes the user want to add or delete.}
}}

