<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GeneDB1}
\alias{GeneDB1}
\title{The Gene specific to the first database class.}
\description{
A gene contains a lot of details that depend on the database it belongs. 
But every gens have information in common. So we put all these informations
into a global class "gene" from which the others specific gene class 
inherit.
}
\section{Slots}{

\describe{
\item{\code{id}}{The id of the gene. Many genes can have the same id because there
are the same in spite of the different attributes they have.}

\item{\code{locus}}{The locus the gene belongs.}

\item{\code{others}}{The others attributes the user want to add or delete.}

\item{\code{genes}}{The list of genes we avec download.}

\item{\code{msU7name}}{an identicator}

\item{\code{fgeneshName}}{an identicator}

\item{\code{rappredname}}{an identicator}

\item{\code{fmin}}{an identicator}

\item{\code{fmax}}{an identicator}

\item{\code{contig}}{an identicator}

\item{\code{iricname}}{an identicator}

\item{\code{strand}}{an identicator}

\item{\code{description}}{a description}
}}

=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GeneDB1}
\alias{GeneDB1}
\title{The Gene specific to the first database class.}
\description{
This gene is specific to the database 1 that is the database "RAPDB". So it
contains all the informations we can have about one gene with the "RAPDB" 
database.
}
\section{Slots}{

\describe{
\item{\code{id}}{The id of the gene. Many genes can have the same id because there
are the same in spite of the different attributes they have.}

\item{\code{locus}}{The locus the gene belongs.}

\item{\code{others}}{The others attributes the user want to add or delete.}

\item{\code{rapDBGeneNameSynonym}}{an identificator}

\item{\code{rapDBGeneSymbolSynonym}}{an identificator}

\item{\code{cgsnlGeneName}}{an identificator}

\item{\code{cgsnlGeneSymbol}}{an identificator}

\item{\code{oryzabaseGeneNameSynonym}}{an identificator}

\item{\code{oryzabaseGeneSymbolSynonym}}{an identificator}

\item{\code{description}}{a description of the gene}
}}

>>>>>>> e1647fc4403c546ffad2654514ec5129355428a1
