% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/methods-GeneDB1.R
\docType{class}
\name{GeneDB1}
\alias{GeneDB1}
\alias{GeneDB1}
\title{The Gene specific to the first database class.}
\usage{
GeneDB1(id, locus, rapDBGeneNameSynonym, rapDBGeneSymbolSynonym, cgsnlGeneName,
  cgsnlGeneSymbol, oryzabaseGeneNameSynonym, oryzabaseGeneSymbolSynonym,
  position, description)
}
\arguments{
\item{id}{The id of the gene. Many genes can have the same id because there
are the same in spite of the different attributes they have.}

\item{locus}{The locus the gene belongs.}

\item{rapDBGeneNameSynonym}{an identificator}

\item{rapDBGeneSymbolSynonym}{an identificator}

\item{cgsnlGeneName}{an identificator}

\item{cgsnlGeneSymbol}{an identificator}

\item{oryzabaseGeneNameSynonym}{an identificator}

\item{oryzabaseGeneSymbolSynonym}{an identificator}

\item{description}{a description of the gene}

\item{others}{The others attributes the user want to add or delete.}
}
\value{
An Experiment
}
\description{
This gene is specific to the database 1 that is the database "RAPDB". So it
contains all the informations we can have about one gene with the "RAPDB"
database.

This constructor allows to create an Experiment from the RAPDB Database.
}
\section{Slots}{

\describe{
\item{\code{id}}{The id of the gene. Many genes can have the same id because there
are the same in spite of the different attributes they have.}

\item{\code{locus}}{The locus the gene belongs.}

\item{\code{others}}{The others attributes the user want to add or delete.}

\item{\code{rapDBGeneNameSynonym}}{an identificator}

\item{\code{rapDBGeneSymbolSynonym}}{an identificator}

\item{\code{cgsnlGeneName}}{an identificator}

\item{\code{cgsnlGeneSymbol}}{an identificator}

\item{\code{oryzabaseGeneNameSynonym}}{an identificator}

\item{\code{oryzabaseGeneSymbolSynonym}}{an identificator}

\item{\code{description}}{a description of the gene}
}}

