% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/constructors-functions.R
\docType{class}
\name{GeneDB3}
\alias{GeneDB3}
\alias{GeneDB3}
\title{An S4 class to represent the third database Gene.}
\usage{
GeneDB3(id, locus, traitGeneId, cgsnlGeneSymbol, GeneSymbolSynonim,
  cgsnlSymbolSynonim, GeneNameSynonim, proteinName, allele, chromosomeNumber,
  explanation, traitClass, rapID, grameneId, arm, locate, geneOntology,
  traitOntology, plantOntology)
}
\arguments{
\item{id}{The id of the gene. Many genes can have the same id because there
are the same in spite of the different attributes they have.}

\item{locus}{The locus the gene belongs.}

\item{traitGeneId}{id}

\item{cgsnlGeneSymbol}{id}

\item{GeneSymbolSynonim}{id}

\item{cgsnlSymbolSynonim}{id}

\item{GeneNameSynonim}{id}

\item{proteinName}{id}

\item{allele}{id}

\item{chromosomeNumber}{id}

\item{explanation}{id}

\item{traitClass}{id}

\item{rapID}{id}

\item{grameneId}{id}

\item{arm}{id}

\item{locate}{id}

\item{geneOntology}{id}

\item{traitOntology}{id}

\item{plantOntology}{id}
}
\value{
An Experiment
}
\description{
This gene is specific to the database 3 that is the database "Oryzabase".
So it contains all the informations we can have about one gene with the
"Oryzabase" database.

This constructor allows to create a GeneDB3 from the DB oryzabase.
}
\section{Slots}{

\describe{
\item{\code{id}}{The id of the gene. Many genes can have the same id because there
are the same in spite of the different attributes they have.}

\item{\code{locus}}{The locus the gene belongs.}

\item{\code{others}}{The others attributes the user want to add or delete.}

\item{\code{traitGeneId}}{id}

\item{\code{cgsnlGeneSymbol}}{id}

\item{\code{GeneSymbolSynonim}}{id}

\item{\code{cgsnlSymbolSynonim}}{id}

\item{\code{GeneNameSynonim}}{id}

\item{\code{proteinName}}{id}

\item{\code{allele}}{id}

\item{\code{chromosomeNumber}}{id}

\item{\code{explanation}}{id}

\item{\code{traitClass}}{id}

\item{\code{rapID}}{id}

\item{\code{grameneId}}{id}

\item{\code{arm}}{id}

\item{\code{locate}}{id}

\item{\code{geneOntology}}{id}

\item{\code{traitOntology}}{id}

\item{\code{plantOntology}}{id}
}}

