% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pratical-functions.R
\name{getAttributes}
\alias{getAttributes}
\title{Function for see the list of the attributes of the class}
\usage{
getAttributes(class)
}
\arguments{
\item{class}{the name of the class you want}
}
\value{
print the attributes of the class
}
\description{
this function will print the attributes of the class
}
\examples{
getAttributes("RAPDB")
}
