% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/get-methods.R
\docType{methods}
\name{getDate}
\alias{getDate}
\alias{getDate,Experiment-method}
\title{Get the date of an object}
\usage{
getDate(object)

\S4method{getDate}{Experiment}(object)
}
\arguments{
\item{object}{The object for which we want to get the name}
}
\value{
The date of the object
}
\description{
Get the date of an object
}
\examples{
exp <- new(Class="Experiment",
 date=Sys.Date(),
  name="ex",
   databases=list(1))
getDate(exp)
}
