% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProperties-function.R
\name{getProperties}
\alias{getProperties}
\title{Get the properties you want from the genes you want}
\usage{
getProperties(geneVector, attributesVector)
}
\arguments{
\item{geneVector}{vector}

\item{attributesVector}{vector}
}
\value{
it will return a dataframe with all the informations you want
}
\description{
Get the properties you want from the genes you want
}
