% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getProperties-methods.R
\docType{methods}
\name{getProperties}
\alias{getProperties}
\alias{getProperties,Experiment-method}
\title{get the properties you want to keep}
\usage{
getProperties(object, idWanted)

\S4method{getProperties}{Experiment}(object, idWanted)
}
\arguments{
\item{object}{The object for which we want to extract a dataframe}

\item{idWanted}{id you want to keep to identify your gene}
}
\value{
a data frame with the genes and the properties you want
}
\description{
get the properties you want to keep
}
