% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/set-methods.R
\docType{methods}
\name{setDate}
\alias{setDate}
\alias{setDate,Experiment-method}
\title{Set the date of an object}
\usage{
setDate(object, date)

\S4method{setDate}{Experiment}(object, date)
}
\arguments{
\item{object}{The object for which we want to set the name}

\item{date}{The date we want to give at the object}
}
\value{
The object with the new name

The object with the new date
}
\description{
Set the date of an object
}
\examples{
exp <- new(Class="Experiment",
 databases=list(1),
 date=Sys.Date(),
 name="test")
 
 
exp <- setDate(exp, "12/18/2010")
}
