% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{updateObject}
\alias{updateObject}
\title{Update an attribute of an object}
\usage{
updateObject(experiment, attribute, value)
}
\arguments{
\item{attribute}{The attribute we want to update}

\item{value}{The value we want to give at the attribute}

\item{object}{The object for which we want modification}
}
\value{
The object without the attribute we want to delete
}
\description{
Update an attribute of an object
}
